+++ pkg/api/v1/groupversion_info.go
@@ -9,7 +9,7 @@ import (
 )
 
 var (
-	// GroupVersion is group version used to register these objects
+	// GroupVersion is the group version used to register these objects
 	GroupVersion = schema.GroupVersion{Group: "k8s.cloudogu.com", Version: "v1"}
 
 	// SchemeBuilder is used to add go types to the GroupVersionKind scheme