+++ pkg/api/v1/restore_types.go
@@ -1,35 +1,42 @@
 package v1
 
 import (
+	"fmt"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"time"
 )
 
 const (
 	RestoreStatusNew        = ""
 	RestoreStatusInProgress = "in progress"
+	RestoreStatusFailed     = "failed"
 	RestoreStatusCompleted  = "completed"
 	RestoreStatusDeleting   = "deleting"
 )
+
+const RestoreFinalizer = "cloudogu-restore-finalizer"
 
 // NOTE: json tags are required.  Any new fields you add must have json tags for the fields to be serialized.
 
 // RestoreSpec defines the desired state of Restore
 type RestoreSpec struct {
-	// INSERT ADDITIONAL SPEC FIELDS - desired state of cluster
-	// Important: Run "make" to regenerate code after modifying this file
-
-	// Foo is an example field of Restore. Edit restore_types.go to remove/update
-	Foo string `json:"foo,omitempty"`
+	// BackupName references the backup that should be restored.
+	BackupName string `json:"backupName,omitempty"`
+	// Provider defines the backup provider which should be used for the restore.
+	Provider Provider `json:"provider,omitempty"`
 }
 
 // RestoreStatus defines the observed state of Restore
 type RestoreStatus struct {
+	// Status represents the state of the backup.
 	Status string `json:"status,omitempty"`
+	// RequeueTimeNanos contains the time in nanoseconds to wait until the next requeue.
+	RequeueTimeNanos time.Duration `json:"requeueTimeNanos,omitempty"`
 }
 
-//+kubebuilder:object:root=true
-//+kubebuilder:subresource:status
-//+kubebuilder:metadata:labels=app=ces;app.kubernetes.io/name=k8s-backup-operator
+// +kubebuilder:object:root=true
+// +kubebuilder:subresource:status
+// +kubebuilder:metadata:labels=app=ces;app.kubernetes.io/name=k8s-backup-operator;k8s.cloudogu.com/part-of=backup
 
 // Restore is the Schema for the restores API
 type Restore struct {
@@ -42,7 +49,7 @@ type Restore struct {
 	Status RestoreStatus `json:"status,omitempty"`
 }
 
-//+kubebuilder:object:root=true
+// +kubebuilder:object:root=true
 
 // RestoreList contains a list of Restore
 type RestoreList struct {
@@ -53,4 +60,24 @@ type RestoreList struct {
 
 func init() {
 	SchemeBuilder.Register(&Restore{}, &RestoreList{})
+}
+
+// GetFieldSelectorWithName return the field selector with the metadata.name field.
+func (r *Restore) GetFieldSelectorWithName() string {
+	return fmt.Sprintf("metadata.name=%s", r.Name)
+}
+
+// GetStatus return the requeueable status.
+func (r *Restore) GetStatus() RequeueableStatus {
+	return r.Status
+}
+
+// GetStatus return the status from the status object.
+func (rs RestoreStatus) GetStatus() string {
+	return rs.Status
+}
+
+// GetRequeueTimeNanos returns the requeue time in nano seconds.
+func (rs RestoreStatus) GetRequeueTimeNanos() time.Duration {
+	return rs.RequeueTimeNanos
 }