+++ CHANGELOG.md
@@ -5,10 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- [#7] Functionality to restore a backup to the namespace where the backup-operator is deployed
+    - Before the restore is applied, resources in this namespace which are irrelevant to the backup process are removed to provide a clean slate
+    - Currently, only the velero provider is supported
 
 ## [v0.3.0] - 2023-10-06
 ### Added
-- [#3] Add functionality to create a backup from the namespace where the backup-operator is deployed. Velero is used as a first provider.
+- [#3] Functionality to create a backup from the namespace where the backup-operator is deployed
+  - Velero is used as a first provider.
 
 ## [v0.2.0] - 2023-10-05
 ### Added